Move poll logic to Entropy_Sources
